moves id extractor to api that uses it
